pyproject.toml
[project]
name = "pdf-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "gradio>=5.47.0",
    "langchain-text-splitters>=0.3.11",
    "markdown>=3.9",
    "numpy>=2.3.3",
    "openai>=1.109.1",
    "pdfminer>=20191125",
    "pdfminer-six>=20250506",
    "pydantic>=2.11.9",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "tomli>=2.2.1",
    "uvicorn>=0.37.0",
]
